router e ed sincronizados. falha no recebimento da resposta de leitura pelo router

+++ src/Lora/Loramesh.h
@@ -59,7 +59,7 @@ typedef struct  {
     uint8_t  devtype;
     uint8_t  devaddr;
     uint8_t  dataslot;
-    uint8_t  seqnum;
+    uint16_t  seqnum;
 } strDevicedescription;
 
 // Fila para armazenar mensagens recebidas
